pyproject.toml
[project]
name = "hot-hog-bicycle"
version = "0.1.0"
description = "Human vs AI hot dog SVG prompt battle — powered by PostHog LLM Analytics"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.50.0",
    "posthog>=3.20.0",
    "sqlalchemy>=2.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]
